I misled @bpilorget by suggesting this.
The confusion was that there is a standard package called just 'comment' https://www.ctan.org/tex-archive/macros/latex/contrib/comment?lang=en but that in older versions included a 'comments.sty' which I thought this could be referring to.  But intead it refers to a style file in the Proposal/ directory.

In the report for D1.2 @nthiery just copied the style file into that directory which happens to work (if the report directory happens to be your working directory).  But other than that if some style is going to be shared between documents a better way is needed (such as a makefile that includes it in the search path).