revert botched merge 4ffa8b5e5c9e213917c79931389a746cfc59b098
this reverted all the unintended changes from that merge
unfortunately, it also reverts *intended* changes from that merge
commit, and it should be redone (with care)

to restore versions of files from the merge commit, it's possible
to run, for example:

    git checkout 4ffa8b5e5c9e213917c79931389a746cfc59b098 -- WP4/D4.11/kernel.tex

and then re-commit that file along with any others

Please remember when merging to with the upstream master to merge
upstream/master *into* your branch, not the other way around. Then
you can push the merged branch to master.