Table of contents
This is taken from commit

commit 6326f54f8f0a253e986f967b1b81773713092394 (HEAD -> pandas, origin/pandas)
Author: Hans Fangohr <fangohr@soton.ac.uk>
Date:   Sun Jun 23 17:18:41 2019 +0200

    add missing pandas dependency

on https://github.com/fangohr/introduction-to-python-for-computational-science-and-engineering (in the pandas branch).

Due to a problem with nbconvert, the pdf generation doesn't currently work on Master.

This is fixed by fixing all the versions in the requirements.txt file in the pandas branch.

I have been told that when nbconvert issues are solved, the pdf creation in the master should work again.